More list to array renames
